drawing triangle instances via buffer offsets
I learned to use a single call to glDrawElementsInstanced(...) to draw
multiple instances of a single triangle with a single offset provided
per instance.  To do this, I bound the float offsets to a buffer object,
specified how to pass them with glVertexAttribPointer(...), and then
restricted the passing of values to 1 (one) value per draw instance with
glVertexAttribDivisor(...).

Cool.