just got the mouse tracking working
it took a fair amount of debugging, but the mouse tracking seems to be
working now.  Apparently it was working, but I had forgotten that I had
unbound a buffer, so when I tried to send updated data to the vertex
shader, there was nothing bound to GL_ARRAY_BUFFER.

Now I need to do some cleanup.